+++ src/attribute/HISTORY.md
@@ -1,6 +1,13 @@
 Attribute Change History
 ========================
 
+@VERSION@
+-----
+
+* Invalid values supplied during Attribute initialization that fail setter
+  validation will now fallback the default value defined in `ATTRS`.
+  [Ticket #2528732] [redbat]
+
 3.8.0
 -----